:sparkles: #260 Feat: Converter 도메인 변경 (Member -> Board)
