+++ src/main/java/kr/co/teacherforboss/converter/BoardConverter.java
@@ -27,6 +27,7 @@ import kr.co.teacherforboss.domain.enums.ImageOrigin;
 import kr.co.teacherforboss.web.dto.BoardRequestDTO;
 import kr.co.teacherforboss.web.dto.BoardResponseDTO;
 import kr.co.teacherforboss.web.dto.HomeResponseDTO;
+import kr.co.teacherforboss.web.dto.MypageResponseDTO;
 import org.springframework.data.domain.Slice;
 
 public class BoardConverter {
@@ -536,6 +537,16 @@ public class BoardConverter {
                 question.getId(), question.getCategory().getName(), question.getTitle(), question.getContent(), question.getAnswerList().size())));
         return HomeResponseDTO.GetHotQuestionsDTO.builder()
                 .hotQuestionList(hotQuestionInfos)
+                .build();
+    }
+
+    public static MypageResponseDTO.GetAnsweredQuestionsDTO toGetAnsweredQuestionsDTO(Slice<Question> questions) {
+        return MypageResponseDTO.GetAnsweredQuestionsDTO.builder()
+                .hasNext(questions.hasNext())
+                .answeredQuestionList(questions.stream().map(question ->
+                        new MypageResponseDTO.GetAnsweredQuestionsDTO.AnsweredQuestion(
+                                question.getTitle(), question.getContent(), question.getCreatedAt(),
+                                question.getCategory().getName(), question.getSolved().isIdentifier(), question.getId())).toList())
                 .build();
     }
 }
